+++ src/Command/Install/CreateVendorsLinksCommand.php
@@ -1,4 +1,5 @@
 <?php
+declare(strict_types=1);
 /**
  * This file is part of me-tools.
  *
@@ -26,22 +27,22 @@ class CreateVendorsLinksCommand extends Command
 {
     /**
      * Hook method for defining this command's option parser
-     * @param ConsoleOptionParser $parser The parser to be defined
-     * @return ConsoleOptionParser
+     * @param \Cake\Console\ConsoleOptionParser $parser The parser to be defined
+     * @return \Cake\Console\ConsoleOptionParser
      */
-    protected function buildOptionParser(ConsoleOptionParser $parser)
+    protected function buildOptionParser(ConsoleOptionParser $parser): ConsoleOptionParser
     {
         return $parser->setDescription(__d('me_tools', 'Creates symbolic links for vendor assets'));
     }
 
     /**
      * Creates symbolic links for vendor assets
-     * @param Arguments $args The command arguments
-     * @param ConsoleIo $io The console io
-     * @return null|int The exit code or null for success
+     * @param \Cake\Console\Arguments $args The command arguments
+     * @param \Cake\Console\ConsoleIo $io The console io
+     * @return int|null The exit code or null for success
      * @uses Command::createLink()
      */
-    public function execute(Arguments $args, ConsoleIo $io)
+    public function execute(Arguments $args, ConsoleIo $io): ?int
     {
         foreach (Configure::read('VENDOR_LINKS') as $origin => $target) {
             $this->createLink(