+++ src/Console/Command.php
@@ -1,4 +1,5 @@
 <?php
+declare(strict_types=1);
 /**
  * This file is part of me-tools.
  *
@@ -27,11 +28,11 @@ abstract class Command extends CakeCommand
     /**
      * Internal method to check if a file already exists and output a warning at
      *  the verbose level
-     * @param ConsoleIo $io The console io
+     * @param \Cake\Console\ConsoleIo $io The console io
      * @param string $path Path
      * @return bool
      */
-    protected function verboseIfFileExists(ConsoleIo $io, $path)
+    protected function verboseIfFileExists(ConsoleIo $io, string $path): bool
     {
         if (!file_exists($path)) {
             return false;
@@ -44,13 +45,13 @@ abstract class Command extends CakeCommand
 
     /**
      * Copies a file
-     * @param ConsoleIo $io The console io
+     * @param \Cake\Console\ConsoleIo $io The console io
      * @param string $source Source file
      * @param string $dest Destination file
      * @return bool
      * @uses verboseIfFileExists()
      */
-    public function copyFile(ConsoleIo $io, $source, $dest)
+    public function copyFile(ConsoleIo $io, string $source, string $dest): bool
     {
         if ($this->verboseIfFileExists($io, $dest)) {
             return false;
@@ -76,13 +77,13 @@ abstract class Command extends CakeCommand
      * Creates a directory.
      *
      * This method creates directories recursively.
-     * @param ConsoleIo $io The console io
+     * @param \Cake\Console\ConsoleIo $io The console io
      * @param string $path Directory path
      * @return bool
      * @uses folderChmod()
      * @uses verboseIfFileExists()
      */
-    public function createDir(ConsoleIo $io, $path)
+    public function createDir(ConsoleIo $io, string $path): bool
     {
         if ($this->verboseIfFileExists($io, $path)) {
             return false;
@@ -102,25 +103,25 @@ abstract class Command extends CakeCommand
 
     /**
      * Creates a file at given path
-     * @param ConsoleIo $io The console io
+     * @param \Cake\Console\ConsoleIo $io The console io
      * @param string $path Where to put the file
      * @param string $contents Content to put in the file
      * @return bool
      * @uses verboseIfFileExists()
      */
-    public function createFile(ConsoleIo $io, $path, $contents)
+    public function createFile(ConsoleIo $io, string $path, string $contents): bool
     {
         return $this->verboseIfFileExists($io, $path) ? false : $io->createFile($path, $contents);
     }
 
     /**
      * Creates a symbolic link
-     * @param ConsoleIo $io The console io
+     * @param \Cake\Console\ConsoleIo $io The console io
      * @param string $source Source file or directory
      * @param string $dest Destination file or directory
      * @return bool
      */
-    public function createLink(ConsoleIo $io, $source, $dest)
+    public function createLink(ConsoleIo $io, string $source, string $dest): bool
     {
         if ($this->verboseIfFileExists($io, $dest)) {
             return false;
@@ -146,12 +147,12 @@ abstract class Command extends CakeCommand
      * Sets folder chmods.
      *
      * This method applies permissions recursively.
-     * @param ConsoleIo $io The console io
+     * @param \Cake\Console\ConsoleIo $io The console io
      * @param string $path Folder path
      * @param int $chmod Chmod
      * @return bool
      */
-    public function folderChmod(ConsoleIo $io, $path, $chmod = 0777)
+    public function folderChmod(ConsoleIo $io, string $path, int $chmod = 0777): bool
     {
         try {
             (new Filesystem())->chmod($path, $chmod, 0000, true);