+++ src/TestSuite/IntegrationTestTrait.php
@@ -1,4 +1,5 @@
 <?php
+declare(strict_types=1);
 /**
  * This file is part of me-tools.
  *
@@ -13,6 +14,8 @@
  */
 namespace MeTools\TestSuite;
 
+use Cake\Controller\Controller;
+use Cake\Event\EventInterface;
 use Cake\TestSuite\Constraint\Session\SessionEquals;
 use Cake\TestSuite\IntegrationTestTrait as CakeIntegrationTestTrait;
 use MeTools\Controller\Component\UploaderComponent;
@@ -28,21 +31,15 @@ trait IntegrationTestTrait
 
     /**
      * Adds additional event spies to the controller/view event manager
-     * @param \Cake\Event\Event $event A dispatcher event
+     * @param \Cake\Event\EventInterface $event A dispatcher event
      * @param \Cake\Controller\Controller|null $controller Controller instance
      * @return void
      */
-    public function controllerSpy($event, $controller = null)
+    public function controllerSpy(EventInterface $event, ?Controller $controller = null): void
     {
         $this->cakeControllerSpy($event, $controller);
 
         $this->_controller->viewBuilder()->setLayout('with_flash');
-
-        //Sets key for cookies
-        if (!$this->_controller->components()->has('Cookie')) {
-            $this->_controller->loadComponent('Cookie');
-        }
-        $this->_controller->Cookie->setConfig('key', 'somerandomhaskeysomerandomhaskey');
 
         if ($this->_controller->components()->has('Uploader')) {
             $this->_controller->Uploader = $this->getMockForComponent(UploaderComponent::class, ['move_uploaded_file']);
@@ -61,7 +58,7 @@ trait IntegrationTestTrait
      *  generated message
      * @return void
      */
-    public function assertCookieIsEmpty($name, $message = '')
+    public function assertCookieIsEmpty(string $name, string $message = ''): void
     {
         $this->_response ?: $this->fail('Not response set, cannot assert cookies');
         $cookie = $this->_response->getCookie($name);
@@ -76,7 +73,7 @@ trait IntegrationTestTrait
      *  generated message
      * @return void
      */
-    public function assertFlashMessage($expected, $key = 0, $message = '')
+    public function assertFlashMessage(string $expected, int $key = 0, string $message = ''): void
     {
         $this->assertSession($expected, sprintf('Flash.flash.%d.message', $key), $message);
     }
@@ -88,7 +85,7 @@ trait IntegrationTestTrait
      *  generated message
      * @return void
      */
-    public function assertResponseOkAndNotEmpty($message = '')
+    public function assertResponseOkAndNotEmpty(string $message = ''): void
     {
         $this->assertResponseOk($message) && $this->assertResponseNotEmpty($message);
     }
@@ -100,7 +97,7 @@ trait IntegrationTestTrait
      * @return void
      * @since 2.18.9
      */
-    public function assertSessionEmpty($path, $message = '')
+    public function assertSessionEmpty(string $path, string $message = ''): void
     {
         $verboseMessage = $this->extractVerboseMessage($message);
         $this->assertThat(null, new SessionEquals($this->_requestSession, $path), $verboseMessage);