+++ tests/TestCase/View/Helper/LibraryHelperTest.php
@@ -1,4 +1,5 @@
 <?php
+declare(strict_types=1);
 /**
  * This file is part of me-tools.
  *
@@ -25,7 +26,7 @@ use MeTools\View\Helper\HtmlHelper;
  */
 class LibraryHelperTest extends HelperTestCase
 {
-    const EXPECTED_DATEPICKER_ICONS = [
+    protected const EXPECTED_DATEPICKER_ICONS = [
         'time' => 'fas fa-clock',
         'date' => 'fas fa-calendar',
         'up' => 'fas fa-chevron-up',
@@ -41,7 +42,7 @@ class LibraryHelperTest extends HelperTestCase
      * Called after every test method
      * @return void
      */
-    public function tearDown()
+    public function tearDown(): void
     {
         parent::tearDown();
 
@@ -79,7 +80,7 @@ class LibraryHelperTest extends HelperTestCase
      */
     public function testBeforeLayout()
     {
-        $this->Helper->beforeLayout(new Event(null), null);
+        $this->Helper->beforeLayout(new Event('myEvent'), null);
         $this->assertEmpty($this->getProperty($this->Helper, 'output'));
         $this->assertEmpty($this->Helper->getView()->fetch('script_bottom'));
 
@@ -90,7 +91,7 @@ class LibraryHelperTest extends HelperTestCase
             '});</script>',
         ];
         $this->setProperty($this->Helper, 'output', ['//first', '//second']);
-        $this->Helper->beforeLayout(new Event(null), null);
+        $this->Helper->beforeLayout(new Event('myEvent'), null);
         $this->assertEmpty($this->getProperty($this->Helper, 'output'));
         $this->assertEquals($expected, preg_split('/' . PHP_EOL . '/', $this->Helper->getView()->fetch('script_bottom')));
     }