+++ tests/test_app/TestApp/Model/Entity/Post.php
@@ -1,4 +1,5 @@
 <?php
+declare(strict_types=1);
 /**
  * This file is part of cakephp-assets.
  *
@@ -17,5 +18,4 @@ use Cake\ORM\Entity;
 
 class Post extends Entity
 {
-
 }